Had to move the packetBuilder code into the main function.
